backend/requirements.txt
twilio
google-auth-oauthlib
google-api-python-client
fastapi
pydantic
python-dotenv
uvicorn
mangum 
firebase-admin
python-multipart
